moved fmt utils to its own file
