+++ src/main/kotlin/org/xodium/vanillaplus/modules/AutoRestartModule.kt
@@ -9,10 +9,10 @@ import net.kyori.adventure.bossbar.BossBar
 import org.xodium.vanillaplus.Config
 import org.xodium.vanillaplus.VanillaPlus.Companion.instance
 import org.xodium.vanillaplus.interfaces.ModuleInterface
+import org.xodium.vanillaplus.utils.FmtUtils.mm
 import org.xodium.vanillaplus.utils.TimeUtils.minutes
 import org.xodium.vanillaplus.utils.TimeUtils.seconds
 import org.xodium.vanillaplus.utils.TimeUtils.ticks
-import org.xodium.vanillaplus.utils.Utils.mm
 import java.time.LocalTime
 import java.time.temporal.ChronoUnit
