+++ src/main/kotlin/org/xodium/vanillaplus/modules/BooksModule.kt
@@ -13,9 +13,9 @@ import org.bukkit.entity.Player
 import org.xodium.vanillaplus.Config
 import org.xodium.vanillaplus.Perms
 import org.xodium.vanillaplus.interfaces.ModuleInterface
+import org.xodium.vanillaplus.utils.FmtUtils.fireFmt
+import org.xodium.vanillaplus.utils.FmtUtils.mm
 import org.xodium.vanillaplus.utils.Utils
-import org.xodium.vanillaplus.utils.Utils.fireFmt
-import org.xodium.vanillaplus.utils.Utils.mm
 
 /**
  * A module that provides a book with rules for players.
