+++ src/main/kotlin/org/xodium/vanillaplus/modules/BroadcastModule.kt
@@ -8,7 +8,7 @@ package org.xodium.vanillaplus.modules
 import org.xodium.vanillaplus.Config
 import org.xodium.vanillaplus.VanillaPlus.Companion.instance
 import org.xodium.vanillaplus.interfaces.ModuleInterface
-import org.xodium.vanillaplus.utils.Utils.mm
+import org.xodium.vanillaplus.utils.FmtUtils.mm
 
 /**
  * A module that broadcasts a random message to all online players at regular intervals.
