+++ src/main/kotlin/org/xodium/vanillaplus/modules/MotdModule.kt
@@ -10,7 +10,7 @@ import org.bukkit.event.EventPriority
 import org.bukkit.event.server.ServerListPingEvent
 import org.xodium.vanillaplus.Config
 import org.xodium.vanillaplus.interfaces.ModuleInterface
-import org.xodium.vanillaplus.utils.Utils.mm
+import org.xodium.vanillaplus.utils.FmtUtils.mm
 
 /**
  * Customizes the server's Message of the Day (MOTD) that appears in the server list.
