+++ src/main/kotlin/org/xodium/vanillaplus/modules/TabListModule.kt
@@ -15,9 +15,9 @@ import org.bukkit.event.player.PlayerJoinEvent
 import org.xodium.vanillaplus.Config
 import org.xodium.vanillaplus.VanillaPlus.Companion.instance
 import org.xodium.vanillaplus.interfaces.ModuleInterface
+import org.xodium.vanillaplus.utils.FmtUtils.mm
 import org.xodium.vanillaplus.utils.TimeUtils.seconds
 import org.xodium.vanillaplus.utils.TimeUtils.ticks
-import org.xodium.vanillaplus.utils.Utils.mm
 
 /**
  * Handles the tab list for all players
